Remove comparison warning from vlsv_reader.cpp

+++ vlsv_reader.cpp
@@ -227,7 +227,7 @@ namespace vlsv {
 
       // Read footer XML tree:
       filein.seekg(footerOffset);
-      if (filein.tellg() != footerOffset) {
+      if (filein.tellg() != (int)footerOffset) {
          lastErrorCode = error::READ_NO_FOOTER;
          success = false;
       }
